Implemented the HttpClient reuse suggestion from https://github.com/polygon-io/client-jvm/issues/12 

I have mixed feelings about it though. 

The positives:
- improved single threaded blocking operations latencies. Fetching historical trades went down for 3-4 seconds to about 1 second
- uses less sockets. Fetching data asynchronously with 3-500 async calls used 5-6 times more connections (according to netstat) before using the patch in this PR

The negatives:
- is much more sensitive to resource contention. The amount of connection related errors tends to increase a lot after making continuous calls for 5-10 minutes. The same happens with the old code as well, but (subjectively) after a longer time. (Side note: the API design makes it hard to differentiate between a "normal" error and a temporary one. We're getting the same ktor exception for a 404 status code, which can be returned when we asked for historical data for a holiday, and also for a socket timeout or some other problem)
- exposes problems with the HttpEngine configurations. OkHttp is most likely configured with 3 or 5 max connections per host, so it's impossible to achieve higher parallelism. ktor doesn't expose that configuration for OkHttp as far as I could see, so this makes OkHttp impossible to use for anything other than toy projects


I've tested OkHttp, CIO, Apache HttpClient before and after the change. 
All were surprisingly better performing for async workloads with the old code, even if I tweaked the maximum number of threads/connections to make sure there's no contention.
All were better for blocking workloads with the new code. (including multithreaded cases).