@mmoghaddam385 2021-03-31
Hi @iulian-radu, thanks for putting this together! Could you expand a bit on how you tested the different configurations? I'm surprised that the Apache and CIO engines configured with many allowed max connections would be slower than the current implementation. 

I'm definitely open to changing the default HTTP engine to something other than OkHttp to address your second negatives bullet point.