@iulian-radu 2021-04-03
I had some hacked calls to the tickers and trades APIs and I was inspecting the results manually. 
Today I've spent some time to add some order to the benchmark code and published it here: https://github.com/iulian-radu/polygon-bench

With the new tool it looks like the latencies and throughput are very similar when we're using parallelism (100 parallel requests in this case, either async or using threads):
```
With reused client

Raw OkHttp async - par: 100 took 17710.9413 ms. 56.462272843736436 fetches / second
Mean: 1.5123470000000017 median: 0.764 std dev: 1.5065833805637836
Errors: 0
Raw OkHttp blocking - par: 100 took 17080.9511 ms. 58.544749302631054 fetches / second
Mean: 1.5353359999999996 median: 1.124 std dev: 1.2617849004897783
Errors: 0
Apache client - 1000 threads (async) - par: 100 took 17813.7503 ms. 56.13641053450716 fetches / second
Mean: 1.6207120000000004 median: 0.718 std dev: 1.8262081373863166
Errors: 1
Apache client - 1000 threads (blocking) - par: 100 took 21129.5642 ms. 47.3270527747089 fetches / second
Mean: 1.9615480961923863 median: 1.526 std dev: 1.41810838795607
Errors: 0
CIO client - 1000 threads (async) - par: 100 took 20183.2498 ms. 49.54603495022887 fetches / second
Mean: 1.8094080000000001 median: 1.219 std dev: 1.1744529652293452
Errors: 0
CIO client - 1000 threads (blocking) - par: 100 took 23671.3212 ms. 42.24521274292032 fetches / second
Mean: 2.1411870000000004 median: 1.857 std dev: 1.2462452535640804
Errors: 0
OkHttp client - 1000 threads (async) - par: 100 took 122730.9835 ms. 8.14790178879321 fetches / second
Mean: 11.914975000000007 median: 12.279 std dev: 4.824577729747445
Errors: 0
OkHttp client - 1000 threads (blocking) - par: 100 took 132547.321 ms. 7.544475380230431 fetches / second
Mean: 12.900176999999987 median: 13.075 std dev: 4.174257953178147
Errors: 0




without reused client

Raw OkHttp async - par: 100 took 24973.283801 ms. 40.04279164760692 fetches / second
Mean: 2.1404249999999996 median: 1.73 std dev: 1.4670000901073583
Errors: 0
Raw OkHttp blocking - par: 100 took 17158.7884 ms. 58.27917313788892 fetches / second
Mean: 1.5298369999999992 median: 0.97 std dev: 1.3925925629670013
Errors: 0
Apache client - 1000 threads (async) - par: 100 took 26492.4579 ms. 37.74659202157305 fetches / second
Mean: 2.4464939999999986 median: 2.1 std dev: 1.5909135934940029
Errors: 1
Apache client - 1000 threads (blocking) - par: 100 took 21959.6248 ms. 45.538118665852615 fetches / second
Mean: 2.015832000000001 median: 1.473 std dev: 1.3477951230717518
Errors: 0
CIO client - 1000 threads (async) - par: 100 took 22510.3501 ms. 44.42400920277113 fetches / second
Mean: 2.1768690000000017 median: 1.457 std dev: 1.5175504748900464
Errors: 0
CIO client - 1000 threads (blocking) - par: 100 took 26138.4042 ms. 38.25788262926931 fetches / second
Mean: 2.3225950000000006 median: 1.988 std dev: 1.3621558989245683
Errors: 0
OkHttp client - 1000 threads (async) - par: 100 took 122340.949 ms. 8.173878069230932 fetches / second
Mean: 11.660769999999983 median: 11.98 std dev: 4.42782199135196
Errors: 0
OkHttp client - 1000 threads (blocking) - par: 100 took 136408.7685 ms. 7.330907030364401 fetches / second
Mean: 12.900830000000006 median: 13.02 std dev: 5.107144382636935
Errors: 0
```

You can switch the PR on and off by commenting/uncommenting lines 30 and 32 here: https://github.com/iulian-radu/polygon-bench/blob/master/app/build.gradle#L30

adjust the level of parallelism here: https://github.com/iulian-radu/polygon-bench/blob/master/app/src/main/java/org/radu/App.java#L65

and enable/disable different configurations here: https://github.com/iulian-radu/polygon-bench/blob/master/app/src/main/java/org/radu/App.java#L75

I've added a custom client that calls OkHttp internally (doesn't deserialize the data though), which seems to be the fastest. 
Out of what `ktor` provides out of the box, the Apache and CIO look decent, while OkHttp is the worst by far due to the host parallelism misconfiguration. The Apache ktor implementation encounters this problem sporadically though: https://github.com/ktorio/ktor/issues/1815

Looks like there's no performance regression after all.