@mmoghaddam385 2021-04-05
Thanks again for the very thorough response! Based on your results, it seems to me like we should keep the reused client implementation and switch `DefaultJvmHttpClientProvider` to use the CIO engine for now. That bug in the apache client is very weird. It seems like there's a fix for it in a newer version of ktor, but I'd like to spend some more time understanding the changes that come with the version upgrades, so we can do that later.

If that all sounds good to you, please go ahead and swap out OkHttp with CIO in `DefaultHttpClientProvider` in this PR and we can get it merged then release a new version. We should continue to support `DefaultOkHttpClientProvider`, but it would be good to add a note about poor parallelism performance in the comment.