@justinpolygon 2023-12-11
Hello, I'm going to close this out since we have significantly changed the way things work since 2021. In addition to updating Ktor. So, I'm going to close this and the corresponding issue https://github.com/polygon-io/client-jvm/issues/12 and we'll take a another look to see if we can improve things.